Add `cli/` package and slim `inference.py` to legacy entry
Completes 8fb0610: `cli/eval/run.py` owns the `run` command + runtime (`main`, `Driver`, helpers);
`inference.py` imports them as the legacy `positronic-inference` entry. No import cycle, no lazy import.
