+++ positronic/cli/eval/__init__.py
@@ -1,0 +1,4 @@
+from positronic.cli.eval.run import run
+
+# Subcommands of `positronic eval`.
+commands = {'run': run}
