+++ positronic/inference.py
@@ -1,8 +1,6 @@
-import logging
+"""Legacy ``positronic-inference`` CLI: the attended ``real``/``phail`` aliases over the ``cli.eval.run`` runner."""
+
 from collections import Counter
-from collections.abc import Callable
-from contextlib import nullcontext
-from dataclasses import dataclass
 from pathlib import Path
 
 import configuronic as cfn
@@ -11,18 +9,14 @@ import pos3
 import pimm
 import positronic.cfg.embodiment
 import positronic.cfg.policy as policy_cfg
-from positronic import utils, wire
-from positronic.dataset.ds_writer_agent import TimeMode
-from positronic.dataset.local_dataset import LocalDatasetWriter, load_all_datasets
-from positronic.eval import Embodiment, Task
+from positronic.cfg.eval.sim.positronic import stack_cubes
+from positronic.cli.eval.run import Driver, main, run
+from positronic.dataset.local_dataset import load_all_datasets
 from positronic.gui.dpg import DearpyguiUi
 from positronic.gui.eval import EvalUI
 from positronic.gui.keyboard import KeyboardControl
-from positronic.policy.base import SampledPolicy
-from positronic.policy.harness import Directive, Harness, default_wrappers
+from positronic.policy.harness import Directive
 from positronic.utils.logging import init_logging
-
-logger = logging.getLogger(__name__)
 
 
 class KeyboardHandler:
@@ -38,20 +32,6 @@ class KeyboardHandler:
             case 'r':
                 return Directive.HOME()
         return None
-
-
-@dataclass
-class Driver:
-    """An attended operator surface: the directive source ``main`` wires into the Harness.
-
-    Driver configs produce a factory called with the resolved local output directory, since
-    the directory exists only after ``pos3.sync`` inside ``main``.
-    """
-
-    gui: DearpyguiUi | None
-    directives: pimm.SignalEmitter
-    directive_wrapper: Callable
-    control_systems: list[pimm.ControlSystem]
 
 
 @cfn.config(ui_scale=1)
@@ -79,104 +59,17 @@ def keyboard(show_gui, task):
     return make
 
 
-def _seed_counter(policy, output_dir: Path):
-    """If policy is a SampledPolicy, seed its episode counter from existing episodes in output_dir."""
-    if not isinstance(policy, SampledPolicy):
-        return
-    try:
-        dataset = load_all_datasets(output_dir)
-    except ValueError:
-        return
-    if len(dataset) == 0:
-        return
-    seeded = policy.counter.seed_from(dataset)
-    logger.info(f'Seeded counter from {seeded} existing episodes')
-
-
-def _completion_sink(policy):
-    """Harness ``on_episode_complete`` callback that tallies completed episodes.
-
-    Returns the ``SampledPolicy``'s counter ``record`` (which reads the sampled
-    key from the session and bumps its tally), or ``None`` for non-sampled
-    policies. The harness fires it on each clean episode completion.
-    """
-    return policy.counter.record if isinstance(policy, SampledPolicy) else None
-
-
-def main(
-    embodiment: Embodiment,
-    policy,
-    driver: Callable[[Path | None], Driver] | None = None,
-    output_dir: str | Path | None = None,
-    task: Task | None = None,
-    trials: list[dict] | None = None,
-    show_gui: bool = False,
-    wrap=default_wrappers,
-):
-    """Run inference for an embodiment, real or simulated.
-
-    ``task`` (when given) supplies the policy-facing instruction, the per-trial ``timeout``
-    bounding self-driven trials, the privileged ground-truth signals to record, and the seeded
-    scene reset run at each trial start; without it the instruction rides the driver. Exactly
-    one of ``driver`` (attended: a factory producing the operator surface that emits the
-    directives) or ``trials`` (unattended: the harness runs the plan itself) must be given;
-    ``show_gui`` applies to the unattended path (attended surfaces bring their own).
-    """
-    assert (driver is None) != (trials is None), 'Provide exactly one of driver or trials'
-    harness = Harness(
-        policy, embodiment, task=task, trials=trials, wrap=wrap, on_episode_complete=_completion_sink(policy)
-    )
-
-    if output_dir is not None:
-        output_dir = pos3.sync(output_dir, sync_on_error=True)
-        utils.save_run_metadata(output_dir, patterns=['*.py', '*.toml'])
-        _seed_counter(policy, output_dir)
-
-    driver = driver(output_dir) if driver is not None else None
-    gui = driver.gui if driver is not None else (DearpyguiUi() if show_gui else None)
-
-    time_mode = TimeMode.MESSAGE if embodiment.simulated else TimeMode.CLOCK
-    writer_cm = LocalDatasetWriter(output_dir) if output_dir is not None else nullcontext(None)
-    with writer_cm as dataset_writer, pimm.World(virtual_time=embodiment.simulated) as world:
-        privileged = task.privileged if task is not None else {}
-        ds_agent = wire.wire_embodiment(world, harness, embodiment, dataset_writer, time_mode, privileged=privileged)
-        if gui is not None:
-            # HACK: GUI cameras are matched to observations by the `image.` name prefix, which
-            # hard-binds GUI wiring to the observation naming convention. TODO: resolve this
-            # coupling (the right binding is still open).
-            for name, obs in embodiment.observations.items():
-                if name.startswith('image.'):
-                    world.connect(obs.source, gui.cameras[name])
-        if driver is not None:
-            world.connect(driver.directives, harness.directive, emitter_wrapper=driver.directive_wrapper)
-        if ds_agent is not None:
-            world.connect(harness.ds_command, ds_agent.command)
-
-        # Sim runs devices + recorder in-process under the virtual clock; real runs them as
-        # background subprocesses. The harness, driver, and GUI placement is identical.
-        foreground = driver.control_systems if driver is not None else []
-        devices = [cs for cs in [*embodiment.control_systems, ds_agent] if cs is not None]
-        if embodiment.simulated:
-            world.run([harness, *foreground, *devices], gui)
-        else:
-            world.run([harness, *foreground], [*devices, gui])
-
-
 run_cfg = cfn.Config(main, embodiment=positronic.cfg.embodiment.droid, policy=policy_cfg.placeholder, driver=keyboard)
 
 
 # Console entry point for [project.scripts].
 @pos3.with_mirror()
 def _internal_main():
-    # Imported here to break the circular import: positronic.cfg.eval.run imports this module.
-    from positronic.cfg.eval.run import run as eval_run  # noqa
-    from positronic.cfg.eval.sim.positronic import stack_cubes  # noqa
-
     init_logging()
     cfn.cli({
         'run': run_cfg,
         'real': run_cfg,  # `real` is the documented name for the hardware path
-        'sim': eval_run.override(eval=stack_cubes),
+        'sim': run.override(eval=stack_cubes),
         'phail': run_cfg.override(
             policy=policy_cfg.phail_multiple,
             driver=eval_ui,
